rearrange functions to avoid dependencies on operator within LAM
